% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGrdTbl.R
\name{createGridTbl}
\alias{createGridTbl}
\title{Create a \pkg{tibble}/dataframe by expanding a set of grids}
\usage{
createGridTbl(...)
}
\arguments{
\item{...}{- named list of 1D grids or name/grid pairs}
}
\value{
\pkg{tibble} object of class 'tbl_df' with outer cross of all 1D grids
}
\description{
Function to create a tibble/dataframe by expanding a set of grids.
}
\details{
The string "grd_" is stripped from any column name in the output
 dataframe.
}
