% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.check.plots.R
\name{gam.check.plots}
\alias{gam.check.plots}
\title{Diagnostic plots for a [mgcv::gam()] output object}
\usage{
gam.check.plots(
  b,
  type = c("deviance", "pearson", "response"),
  rep = 0,
  level = 0.9,
  s.rep = 10,
  combinePlots = TRUE,
  ...
)
}
\arguments{
\item{b}{- \pkg{mgcv} gam object}

\item{type}{- type of residuals ("deviance","pearson","response")}

\item{rep}{- (default = 0)}

\item{level}{- (default = 0.9)}

\item{s.rep}{- (default = 10)}

\item{combinePlots}{- flag (T/F) to combine plots (T) or return a list (F)}

\item{...}{- not used}
}
\value{
ggplot2 plot object (combinePlots=TRUE) or list of such (combinePlots=FALSE)
}
\description{
Function to make diagnostic plots for a [mgcv::gam()] output object.
}
\details{
This function basically creates \pkg{ggplot2} versions of the plots provided by 
[mgcv::gam.check()].
}
