% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.check.R
\name{gam.check}
\alias{gam.check}
\title{Check a [mgcv::gam()] model (printout only)}
\usage{
gam.check(
  b,
  type = c("deviance", "pearson", "response"),
  k.sample = 5000,
  k.rep = 200
)
}
\arguments{
\item{b}{- a fitted gam object produced by [mgcv::gam()]}

\item{type}{- residuals type (see mgcv)}

\item{k.sample}{- above this k uses random sub-sample of data}

\item{k.rep}{number of reshuffles to do to get p-value for k-testing}
}
\value{
character vector of output printed by [mgcv::gam.check()]
}
\description{
Function to check a [mgcv::gam()] model (printout only). A port 
of [mgcv::gam.check()] that does not include the diagnostic plots.
}
\details{
This function is a re-working of [mgcv::gam.check()] to provide 
a character vector output suitable for using, e.g., in a markdown document.
}
