% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmoothTerms.R
\name{getSmoothTerms}
\alias{getSmoothTerms}
\title{Get the names of the smooth terms in a [mgcv::gam()] model}
\usage{
getSmoothTerms(model)
}
\arguments{
\item{model}{- the [mgcv::gam()] gam model}
}
\value{
a character vector with the names of the smooth terms
}
\description{
Function to get the names of the smooth terms in a gam model.
}
