% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarsForSmoothTerms.R
\name{getVarsForSmoothTerms}
\alias{getVarsForSmoothTerms}
\title{Get the covariate names for the smooth terms in a [mgcv::gam()] model}
\usage{
getVarsForSmoothTerms(terms)
}
\arguments{
\item{terms}{- character vector with the names of smooth terms from a [mgcv::gam()] gam model}
}
\value{
a named list with a character vector with the names of the covariates for each smooth term
}
\description{
Function to get the covariate names for the smooth terms in a gam model.
}
\details{
The names of the list elements are the terms for which the covariate names were determined.
}
