% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmoothTerm1D.R
\name{plotSmoothTerm1D}
\alias{plotSmoothTerm1D}
\title{Plot a [mgcv::gam()] model smooth 1D term}
\usage{
plotSmoothTerm1D(term, lstPredTermTbls, dfrDat, xlab = NULL, ci = 0.8)
}
\arguments{
\item{term}{- name of smooth 1D term to plot}

\item{lstPredTermTbls}{- list of dataframes with predicted values for terms (or single dataframe)}

\item{dfrDat}{- original dataframe for gam model (for rug plots) or NULL (for no rug plots)}

\item{xlab}{- label for x axis}

\item{ci}{- tow-sided confidence interval to plot}
}
\value{
ggplot2 plot object of class gg
}
\description{
Function to plot a [mgcv::gam()] model 1D smooth term.
}
\details{
lstPredTermTbls should be a list (of dataframes) from running [predSmoothTerms()].
}
