% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmoothTerm2D.R
\name{plotSmoothTerm2D}
\alias{plotSmoothTerm2D}
\title{Plot a [mgcv::gam()] model smooth 2D term}
\usage{
plotSmoothTerm2D(
  term,
  lstPredTermTbls,
  dfrDat = NULL,
  xlab = NULL,
  ylab = NULL,
  ori = "H"
)
}
\arguments{
\item{term}{- name of smooth 2D term to plot}

\item{lstPredTermTbls}{- list of dataframes with predicted values for terms (or single dataframe)}

\item{dfrDat}{- original dataframe for gam model (for rug plots) or NULL (for no rug plots)}

\item{xlab}{- label for x axis}

\item{ori}{- orientation for stacking the value/se plots ('H' or 'V')}

\item{yLab}{- label for y axis}
}
\value{
ggplot2 plot object of class gg
}
\description{
Function to plot a [mgcv::gam()] model 2D smooth term.
}
\details{
lstPredTermTbls should be a list (of dataframes) from running [predSmoothTerms()].
}
