% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmoothTerms.R
\name{plotSmoothTerms}
\alias{plotSmoothTerms}
\title{Plot all [mgcv::gam()] model smooth terms}
\usage{
plotSmoothTerms(lstPredTermTbls, labs, dfrDat = NULL, ori = "H", ci = 0.8)
}
\arguments{
\item{lstPredTermTbls}{- list of dataframes with predicted values for terms (or single dataframe)}

\item{labs}{- list of axis labels for covariates}

\item{dfrDat}{- original dataframe for gam model (for rug plots) or NULL (for no rug plots)}

\item{ori}{- orientation for stacking the 2D value/se plots ('H' or 'V')}
}
\value{
list of ggplot2 plot objects (all of class gg)
}
\description{
Function to plot all [mgcv::gam()] model smooth term.
}
\details{
\code{lstPredTermTbls} should be a list (of dataframes) from running [predSmoothTerms()]. 
\code{labs} should be a named list of 1-element character vectors, with names corresponding to 
term names and values the associated label to use. Currently, only 1D and 2D smooth terms can be 
plotted.

The intercept is added to the first model term, but not to subsequent terms.
}
