% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSmoothTerm.R
\name{predSmoothTerm}
\alias{predSmoothTerm}
\title{Predict values for a smooth term}
\usage{
predSmoothTerm(mdl, dfr, term, include_intercept = FALSE)
}
\arguments{
\item{dfr}{- dataframe with covariates at which to predict the smooth term}

\item{term}{- name of the smooth term to do prediction for}

\item{model}{- [mgcv::gam()] model object}

\item{intercept}{- flag (T/F) to include intercept in returned values}
}
\value{
tibble dataframe (tbl_df object)
}
\description{
Function to predict values for a smooth term
}
