% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.gam.R
\name{qq.gam}
\alias{qq.gam}
\title{QQ plot for [mgcv::gam()] output object}
\usage{
qq.gam(
  object,
  rep = 0,
  level = 0.9,
  s.rep = 10,
  type = c("deviance", "pearson", "response"),
  ...
)
}
\arguments{
\item{object}{- gam object}

\item{rep}{- (default = 0)}

\item{level}{- (default = 0.9)}

\item{s.rep}{- (default = 10)}

\item{type}{- type of residuals ("deviance","pearson","response")}

\item{...}{- not used}
}
\value{
ggplot2 plot object
}
\description{
Function to make a QQ plot for [mgcv::gam()] output object.
}
\details{
A ggplot2 version of [mgcv::qq.gam()].
}
